refactoring in the test structure

+++ src/azureDevOps/pipeline/pipeline.tests.ps1
@@ -46,23 +46,25 @@ Describe "YAML Pipelines" -Tag dotnetCore {
 
 AfterAll {
 
-    Write-Host "Test execution finished. Tearing down pipelines." -ForegroundColor Yellow
-    $pipelines = Get-AzureDevOpsPipelines `
-        -personalAccessToken $env:personalAccessToken `
-        -orgUrl $env:orgUrl `
-        -teamProject $env:testsTeamProject
-
-    Write-Host "Removing ALL pipelines on Team Project $teamProject" -ForegroundColor Yellow
-    
-    foreach ($pipeline in $pipelines) {
-        Write-Host "Removing pipeline:" -ForegroundColor Yellow
-        Write-Host "Name: $($pipeline.name)" -ForegroundColor Yellow
-        Write-Host "Id $($pipeline.id)" -ForegroundColor Yellow
-
-        Remove-AzureDevOpsPipelines `
+    if (-not $env:skipTearDown) {
+        Write-Host "Test execution finished. Tearing down pipelines." -ForegroundColor Yellow
+        $pipelines = Get-AzureDevOpsPipelines `
             -personalAccessToken $env:personalAccessToken `
             -orgUrl $env:orgUrl `
-            -teamProject $env:testsTeamProject `
-            -pipelineId  $pipeline.id
-    }   
+            -teamProject $env:testsTeamProject
+
+        Write-Host "Removing ALL pipelines on Team Project $teamProject" -ForegroundColor Yellow
+    
+        foreach ($pipeline in $pipelines) {
+            Write-Host "Removing pipeline:" -ForegroundColor Yellow
+            Write-Host "Name: $($pipeline.name)" -ForegroundColor Yellow
+            Write-Host "Id $($pipeline.id)" -ForegroundColor Yellow
+
+            Remove-AzureDevOpsPipelines `
+                -personalAccessToken $env:personalAccessToken `
+                -orgUrl $env:orgUrl `
+                -teamProject $env:testsTeamProject `
+                -pipelineId  $pipeline.id
+        }   
+    }
 }
